Update dependencies
    $ ./reinstall_packages.sh

The `package.json` files have loose version dependencies.  If we run
`npm update`, it will replace these loose dependencies with exact
dependencies to the latest version.

If, instead, we remove `node_modules` and run `npm install`, then it
leaves `package.json` as it is and simply pulls the most recent version
of these packages as it regenerates `node_modules`.